pyproject.toml
[project]
name = "ai-trip-planner"
version = "0.1.0"
description = "AI Agent for travel planning & flight monitoring"
author = [
    {name = "Janet Zhai", email = "janet.zhai1001@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=0.0.8",
    "langchain-community>=0.0.20",
    "streamlit>=1.31.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "requests>=2.31.0",
    "pandas>=2.2.0",
    "google-search-results>=2.4.2",
    "langgraph>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]
